added static file

+++ new_pro/shorten/views.py
@@ -11,7 +11,7 @@ from django.urls import reverse
 from django.contrib.auth.decorators import login_required
 
 def index(request):
-    objects= bitly.objects.all()[: :-1]
+    objects= bitly.objects.filter(user__iexact=request.user.username)[: :-1]
     print(objects)
 
     context={'objs': objects}
@@ -23,6 +23,7 @@ def create(request):
     if form.is_valid():
         instance = form.save(commit= False)
         instance.shortcode = create_shortcode()
+        instance.user=request.user.username
         instance.datewise = "{}"
         instance.save()
 
@@ -47,19 +48,21 @@ def goto(request,shortcode=None):
 def update(request,pk=None):
     if request.user.is_authenticated:
         qs=get_object_or_404(bitly,id=pk)
-        form=editBitly(request.POST or None,instance=qs)
+        if qs and request.user.username==qs.user:
+            form=editBitly(request.POST or None,instance=qs)
 
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect(reverse('index'))
-        context={'urlform':form}
-        return render(request,"create.html",context)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect(reverse('index'))
+            context={'urlform':form}
+            return render(request,"create.html",context)
     return HttpResponseRedirect(reverse('index'))
 
 def delete(request,pk=None):
     if request.user.is_authenticated:
         qs=get_object_or_404(bitly,id=pk)
-        qs.delete()
+        if qs and request.user.username==qs.user:
+            qs.delete()
     return HttpResponseRedirect(reverse('index'))
